+++ apps/modernization-ui/src/design-system/card/collapsible/collapsible-card.module.scss
@@ -26,8 +26,6 @@
         display: flex;
         justify-content: space-between;
         align-items: center;
-        font-weight: 700;
-        font-size: 1.375rem;
         @extend %thin-bottom;
     }
